Rename home package
